evaluation/requirements.txt
# Azure AI Evaluation SDK
azure-ai-evaluation>=1.0.0

# Required for Azure authentication
azure-identity>=1.19.0

# OpenTelemetry for tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0

# Data processing
pandas>=2.0.0
